pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-takt"
version = "0.1.40"
description = "Multi-agent orchestration for AI-assisted software development"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6,<7",
    "textual>=0.85,<1",
    "onnxruntime>=1.17,<2",
    "sqlite-vec>=0.1,<1",
    "tokenizers>=0.19,<1",
    "numpy>=1.26,<2",
]

[project.scripts]
takt = "agent_takt.cli:main"
takt-fleet = "agent_takt_fleet.cli:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-xdist>=3",
    "pytest-timeout>=2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agent_takt = [
    "_data/templates/agents/*.md",
    "_data/templates/skills/**/*",
    "_data/docs/memory/*.md",
    "_data/default_config.yaml",
    "_data/agents_skills/**/*",
    "_data/claude_skills/**/*",
    "_data/claude_agents/**",
]

[tool.pytest.ini_options]
timeout = 120
timeout_method = "thread"

[tool.uv]
package = true
default-groups = ["dev"]
